+++ htdocs/_assets/css/about-index-1.css
@@ -1,0 +1,1 @@
+:root{--color-txt: #463e2c;--color-reversal: #fff;--color-gray: #d6d4d0;--color-prime: #463e2c;--color-second: red;--color-third: blue;--color-body: #fff;--color-highlight: #f8f8f7;--font-leading: calc(27 / 15);--font-tracking: 0;--font-weight-normal: 400;--font-feature-settings: normal;--transition: 0.4s ease-out;--headerHeight: 7.2rem}@media screen and (width >= 48em){:root{--headerHeight: 10.3rem;--font-tracking: 0;--font-leading: calc(30 / 16)}}:root{--size-container: 1200;--size-logo: 200;--leading-trim: calc((1em - 1lh) / 2);--font-family-base: Noto Sans JP, NotoSansJP, Noto Sans Japanese, sans-serif;--font-setting-base: var(--font-weight-normal) var(--font-size-base) / var(--font-leading) var(--font-family-base);--font-size-heading-1: 4rem;--font-size-heading-2: 3.6rem;--font-size-heading-3: 2.8rem;--font-size-heading-4: 2.4rem;--font-size-heading-5: 2rem;--font-size-large-2: 1.8rem;--font-size-large-1: 1.6rem;--font-size-base: 1.4rem;--font-size-small-1: 1.2rem;--font-size-small-2: 1rem}@media screen and (width >= 48em){:root{--font-size-heading-1: 6.4rem;--font-size-heading-2: 4.4rem;--font-size-heading-3: 3.6rem;--font-size-heading-4: 2.8rem;--font-size-heading-5: 2.4rem;--font-size-large-2: 2rem;--font-size-large-1: 1.8rem;--font-size-base: 1.6rem;--font-size-small-1: 1.4rem;--font-size-small-2: 1.2rem}}.c_ttl_lower[data-astro-cid-k5h775l4]{font-size:var(--font-size-heading-1);text-align:center;font-weight:bold;padding-block:1em}:root{--color-txt: #463e2c;--color-reversal: #fff;--color-gray: #d6d4d0;--color-prime: #463e2c;--color-second: red;--color-third: blue;--color-body: #fff;--color-highlight: #f8f8f7;--font-leading: calc(27 / 15);--font-tracking: 0;--font-weight-normal: 400;--font-feature-settings: normal;--transition: 0.4s ease-out;--headerHeight: 7.2rem}@media screen and (width >= 48em){:root{--headerHeight: 10.3rem;--font-tracking: 0;--font-leading: calc(30 / 16)}}:root{--size-container: 1200;--size-logo: 200;--leading-trim: calc((1em - 1lh) / 2);--font-family-base: Noto Sans JP, NotoSansJP, Noto Sans Japanese, sans-serif;--font-setting-base: var(--font-weight-normal) var(--font-size-base) / var(--font-leading) var(--font-family-base);--font-size-heading-1: 4rem;--font-size-heading-2: 3.6rem;--font-size-heading-3: 2.8rem;--font-size-heading-4: 2.4rem;--font-size-heading-5: 2rem;--font-size-large-2: 1.8rem;--font-size-large-1: 1.6rem;--font-size-base: 1.4rem;--font-size-small-1: 1.2rem;--font-size-small-2: 1rem}@media screen and (width >= 48em){:root{--font-size-heading-1: 6.4rem;--font-size-heading-2: 4.4rem;--font-size-heading-3: 3.6rem;--font-size-heading-4: 2.8rem;--font-size-heading-5: 2.4rem;--font-size-large-2: 2rem;--font-size-large-1: 1.8rem;--font-size-base: 1.6rem;--font-size-small-1: 1.4rem;--font-size-small-2: 1.2rem}}li[data-astro-cid-ilhxcym7]:not(:last-child){position:relative}li[data-astro-cid-ilhxcym7]:not(:last-child)::after{position:absolute;inset:0.1em -2rem 0 auto;height:0.5em;aspect-ratio:1;margin:auto;content:"";border:0.1rem solid var(--color-prime);border-width:0.2rem 0.2rem 0 0;rotate:45deg}li[data-astro-cid-ilhxcym7] a[data-astro-cid-ilhxcym7]{transition:opacity var(--transition)}@media (any-hover: hover){li[data-astro-cid-ilhxcym7] a[data-astro-cid-ilhxcym7]:hover{opacity:0.7}}li[data-astro-cid-ilhxcym7] a[data-astro-cid-ilhxcym7]:focus-visible{opacity:0.7}
