+++ htdocs/_assets/css/index.css
@@ -1,4 +1,4 @@
-@charset "UTF-8";:root{--color-txt: #463e2c;--color-reversal: #fff;--color-gray: #d6d4d0;--color-prime: #463e2c;--color-second: red;--color-third: blue;--color-body: #fff;--color-highlight: #f8f8f7;--font-leading: calc(27 / 15);--font-tracking: 0;--font-weight-normal: 400;--font-feature-settings: normal;--transition: 0.4s ease-out;--headerHeight: 7.2rem}@media screen and (width >= 48em){:root{--headerHeight: 10.3rem;--font-tracking: 0;--font-leading: calc(30 / 16)}}:root{--size-container: 1200;--size-logo: 200;--leading-trim: calc((1em - 1lh) / 2);--font-family-base: Noto Sans JP, NotoSansJP, Noto Sans Japanese, sans-serif;--font-setting-base: var(--font-weight-normal) var(--font-size-base) / var(--font-leading) var(--font-family-base);--font-size-heading-1: 4rem;--font-size-heading-2: 3.6rem;--font-size-heading-3: 2.8rem;--font-size-heading-4: 2.4rem;--font-size-heading-5: 2rem;--font-size-large-2: 1.8rem;--font-size-large-1: 1.6rem;--font-size-base: 1.4rem;--font-size-small-1: 1.2rem;--font-size-small-2: 1rem}@media screen and (width >= 48em){:root{--font-size-heading-1: 6.4rem;--font-size-heading-2: 4.4rem;--font-size-heading-3: 3.6rem;--font-size-heading-4: 2.8rem;--font-size-heading-5: 2.4rem;--font-size-large-2: 2rem;--font-size-large-1: 1.8rem;--font-size-base: 1.6rem;--font-size-small-1: 1.4rem;--font-size-small-2: 1.2rem}}section[data-astro-cid-e6izjgoh]{height:3rem;color:var(--color-second);width:3rem}section[data-astro-cid-e6izjgoh]::after{content:""}@keyframes splide-loading{0%{transform:rotate(0)}to{transform:rotate(1turn)}}.splide__track--draggable{-webkit-touch-callout:none;-webkit-user-select:none;-ms-user-select:none;user-select:none}.splide__track--fade>.splide__list>.splide__slide{margin:0!important;opacity:0;z-index:0}.splide__track--fade>.splide__list>.splide__slide.is-active{opacity:1;z-index:1}.splide--rtl{direction:rtl}.splide__track--ttb>.splide__list{display:block}.splide__container{box-sizing:border-box;position:relative}.splide__list{backface-visibility:hidden;display:-ms-flexbox;display:flex;height:100%;margin:0!important;padding:0!important}.splide.is-initialized:not(.is-active) .splide__list{display:block}.splide__pagination{-ms-flex-align:center;align-items:center;display:-ms-flexbox;display:flex;-ms-flex-wrap:wrap;flex-wrap:wrap;-ms-flex-pack:center;justify-content:center;margin:0;pointer-events:none}.splide__pagination li{display:inline-block;line-height:1;list-style-type:none;margin:0;pointer-events:auto}.splide:not(.is-overflow) .splide__pagination{display:none}.splide__progress__bar{width:0}.splide{position:relative;visibility:hidden}.splide.is-initialized,.splide.is-rendered{visibility:visible}.splide__slide{backface-visibility:hidden;box-sizing:border-box;-ms-flex-negative:0;flex-shrink:0;list-style-type:none!important;margin:0;position:relative}.splide__slide img{vertical-align:bottom}.splide__spinner{animation:splide-loading 1s linear infinite;border:2px solid #999;border-left-color:transparent;border-radius:50%;bottom:0;contain:strict;display:inline-block;height:20px;left:0;margin:auto;position:absolute;right:0;top:0;width:20px}.splide__sr{clip:rect(0 0 0 0);border:0;height:1px;margin:-1px;overflow:hidden;padding:0;position:absolute;width:1px}.splide__toggle.is-active .splide__toggle__play,.splide__toggle__pause{display:none}.splide__toggle.is-active .splide__toggle__pause{display:inline}.splide__track{overflow:hidden;position:relative;z-index:0}
+@charset "UTF-8";@keyframes splide-loading{0%{transform:rotate(0)}to{transform:rotate(1turn)}}.splide__track--draggable{-webkit-touch-callout:none;-webkit-user-select:none;-ms-user-select:none;user-select:none}.splide__track--fade>.splide__list>.splide__slide{margin:0!important;opacity:0;z-index:0}.splide__track--fade>.splide__list>.splide__slide.is-active{opacity:1;z-index:1}.splide--rtl{direction:rtl}.splide__track--ttb>.splide__list{display:block}.splide__container{box-sizing:border-box;position:relative}.splide__list{backface-visibility:hidden;display:-ms-flexbox;display:flex;height:100%;margin:0!important;padding:0!important}.splide.is-initialized:not(.is-active) .splide__list{display:block}.splide__pagination{-ms-flex-align:center;align-items:center;display:-ms-flexbox;display:flex;-ms-flex-wrap:wrap;flex-wrap:wrap;-ms-flex-pack:center;justify-content:center;margin:0;pointer-events:none}.splide__pagination li{display:inline-block;line-height:1;list-style-type:none;margin:0;pointer-events:auto}.splide:not(.is-overflow) .splide__pagination{display:none}.splide__progress__bar{width:0}.splide{position:relative;visibility:hidden}.splide.is-initialized,.splide.is-rendered{visibility:visible}.splide__slide{backface-visibility:hidden;box-sizing:border-box;-ms-flex-negative:0;flex-shrink:0;list-style-type:none!important;margin:0;position:relative}.splide__slide img{vertical-align:bottom}.splide__spinner{animation:splide-loading 1s linear infinite;border:2px solid #999;border-left-color:transparent;border-radius:50%;bottom:0;contain:strict;display:inline-block;height:20px;left:0;margin:auto;position:absolute;right:0;top:0;width:20px}.splide__sr{clip:rect(0 0 0 0);border:0;height:1px;margin:-1px;overflow:hidden;padding:0;position:absolute;width:1px}.splide__toggle.is-active .splide__toggle__play,.splide__toggle__pause{display:none}.splide__toggle.is-active .splide__toggle__pause{display:inline}.splide__track{overflow:hidden;position:relative;z-index:0}
 :root {
   --color-txt: #463e2c;
   --color-reversal: #fff;
@@ -238,7 +238,6 @@
 .c_slide .splide__toggle {
   position: relative;
   display: block;
-  width: min(10.6666666667vw,8rem);
   height: min(10.6666666667vw,8rem);
   aspect-ratio: 1;
   color: transparent;
