+++ htdocs/_assets/js/app.js
@@ -1,187 +1,3 @@
-class Accordion {
-  /**
-   * アコーディオンの開閉スピード
-   * @parm {number}
-   */
-  speed;
-  /**
-   * SPのブレイクポイント
-   * @parm {number}
-   */
-  mq;
-  constructor(SPEED = 250, MQ = 768) {
-    this.speed = SPEED;
-    this.mq = MQ;
-    const OFFSET_TIME = 5;
-    const HEIGHT_CLOSED = "--pullHeightClosed";
-    const HEIGHT = "--pullHeightOpen";
-    const accordions = document.querySelectorAll(".c_pull");
-    accordions.forEach((accordion) => {
-      const ttl = accordion.querySelector(".c_pull_ttl");
-      const content = accordion.querySelector(".c_pull_content");
-      const target = accordion;
-      const hasOpenedClass = accordion.classList.contains("-open");
-      ttl?.addEventListener("click", (e) => {
-        e.preventDefault();
-        const ttlHeight = ttl.offsetHeight;
-        target.style.setProperty(HEIGHT_CLOSED, `${ttlHeight}px`);
-        if (!target.open) {
-          target.open = true;
-          target.style.setProperty(HEIGHT, `${ttlHeight + content.offsetHeight}px`);
-          setTimeout(() => {
-            accordion.classList.add("-open");
-            content?.focus({ preventScroll: true });
-            if (document.activeElement !== content) {
-              content?.setAttribute("tabindex", "-1");
-              content?.focus({ preventScroll: true });
-            }
-          }, OFFSET_TIME);
-        } else if (target.open) {
-          target.style.setProperty(HEIGHT, `${ttlHeight + content.offsetHeight}px`);
-          setTimeout(() => {
-            accordion.classList.remove("-open");
-          }, OFFSET_TIME);
-          setTimeout(() => {
-            target.open = false;
-          }, SPEED + OFFSET_TIME);
-        }
-      });
-      accordion.addEventListener("toggle", () => {
-        if (target.open && !hasOpenedClass) {
-          accordion.classList.add("-open");
-        } else if (!target.open && hasOpenedClass) {
-          accordion.classList.remove("-open");
-        }
-      });
-      function spPull() {
-        const ww = window.innerWidth;
-        if (target.classList.contains("-spPull")) {
-          if (ww <= MQ) {
-            accordion.classList.remove("-open");
-            target.open = false;
-          } else {
-            accordion.classList.add("-open");
-            target.open = true;
-          }
-        }
-      }
-      spPull();
-      window.addEventListener("resize", spPull);
-    });
-  }
-}
-
-const sanitize = (str) => {
-  return String(str).replace(/&/g, "&amp;").replace(/"/g, "&quot;").replace(/</g, "&lt;").replace(/>/g, "&gt;");
-};
-
-class Modal {
-  /**
-   * モーダルのクラス
-   * @parm {string}
-   */
-  target;
-  constructor(TARGET = ".c_modal") {
-    this.target = TARGET;
-    const modalBtn = document.querySelectorAll(TARGET + "_btn");
-    let closes = document.querySelectorAll(TARGET + "_close");
-    modalBtn.forEach((a) => {
-      a.addEventListener("click", (e) => {
-        const event = e.currentTarget;
-        let modalID = sanitize(event?.getAttribute("aria-controls") || "");
-        let modal = document.getElementById(modalID);
-        const targetNext = event.nextElementSibling;
-        if (modalID) {
-          modal = document.getElementById(modalID);
-        } else {
-          const modalSrc = sanitize(event.getAttribute("data-src"));
-          const modalAlt = sanitize(event.getAttribute("data-alt"));
-          if (!targetNext) {
-            const YOUTUBE = /(youtube(-nocookie)?\.com|youtu\.be)\/(watch\?v=|v\/|u\/|embed\/?)?([\w-]{11})(.*)?/i;
-            const youtube_uri = YOUTUBE.exec(modalSrc);
-            let dialog = "";
-            dialog += '<dialog class="c_modal"><div class="c_modal_content" tabindex="-1">';
-            if (youtube_uri) {
-              dialog += '<button class="c_modal_close"><span class="txtHidden">モーダルウィンドウを閉じる</span></button>';
-              dialog += setFrame(
-                appendQueryParams(
-                  "https://www.youtube" + (youtube_uri[2] || "") + ".com/embed/" + youtube_uri[4],
-                  //youtubeのオプションを記述
-                  Object.assign(
-                    {
-                      autoplay: 1,
-                      rel: 0
-                    },
-                    parseQueryParams(youtube_uri[5] || "")
-                  )
-                )
-              );
-            } else {
-              dialog += `<figure><img src=${modalSrc} decoding="async" alt=${modalAlt}></figure>`;
-            }
-            dialog += '<button class="c_modal_close"><span class="txtHidden">モーダルウィンドウを閉じる</span></button>';
-            dialog += "</div></dialog>";
-            event.insertAdjacentHTML("afterend", dialog);
-          }
-          modal = event.nextElementSibling;
-          closes = document.querySelectorAll(".c_modal_close");
-        }
-        modal?.showModal();
-        const content = modal.querySelector(TARGET + "_content");
-        setTimeout(() => {
-          content?.focus({ preventScroll: true });
-          if (document.activeElement !== content) {
-            content?.setAttribute("tabindex", "-1");
-            content?.focus({ preventScroll: true });
-          }
-        }, 0);
-        modalID = null;
-        closes.forEach((close) => {
-          close.addEventListener("click", () => {
-            closeModal(modal);
-          });
-        });
-        modal?.addEventListener("cancel", () => {
-          closeModal(modal);
-        });
-        modal?.addEventListener("click", (e2) => {
-          const event2 = e2.target;
-          if (event2 === modal) {
-            closeModal(modal);
-          }
-        });
-      });
-    });
-    function closeModal(modal) {
-      modal.close("cancelled");
-      if (modal.querySelector(".frameWrapper")) {
-        modal.remove();
-      }
-    }
-    function setFrame(target) {
-      return '<div class="frameWrapper"><iframe frameborder="0" allow="autoplay; fullscreen" src="' + target + '"/></div>';
-    }
-    function parseQueryParams(params) {
-      const pairs = decodeURI(params.split("#")[0]).split("&");
-      const obj = /* @__PURE__ */ new Map();
-      let p;
-      for (let i = 0, n = pairs.length; i < n; i++) {
-        if (!pairs[i]) {
-          continue;
-        }
-        p = pairs[i].split("=");
-        obj.set(p[0], p[1]);
-      }
-      return obj;
-    }
-    function appendQueryParams(url, params) {
-      const keys = Object.keys(params);
-      const query = keys.map((key) => `${key}=${params[key]}`).join("&");
-      return `${url}${url.includes("?") ? "&" : "?"}${query}`;
-    }
-  }
-}
-
 class ScrollAnimation {
   /**
    * アニメーションを付与する要素
@@ -223,57 +39,76 @@ class ScrollAnimation {
   }
 }
 
-class Tab {
+class Hamburger {
   /**
-   * タブのクラス
-   * @parm {string}
+   * ナビゲーション要素のCSSセレクター
+   * @param {string} target - ナビゲーション要素を指定するCSSセレクター
    */
   target;
   /**
-   * タブ展開時のクラス
-   * @parm {string}
+   * メニュー展開時に付与するクラス名
+   * @param {string} open - メニューが開いている状態を示すクラス名
    */
   open;
-  constructor(TARGET = ".c_tab", OPEN = "-open") {
+  /**
+   * ハンバーガーメニューのインスタンスを生成し、イベントリスナーを設定します。
+   * @param {string} [TARGET=".headerNavi"] - ナビゲーション要素を指定するCSSセレクター（デフォルト: ".headerNavi"）
+   * @param {string} [OPEN="-open"] - メニュー展開時に付与するクラス名（デフォルト: "-open"）
+   */
+  constructor(TARGET = ".headerNavi", OPEN = "-open") {
     this.target = TARGET;
     this.open = OPEN;
-    const btns = document.querySelectorAll(TARGET + "_list li button");
-    function onTabClick(e) {
-      const event = e.target;
-      const parent = event.closest(TARGET);
-      const tabContents = parent?.querySelectorAll(TARGET + "_content");
-      const tabArr = Array.prototype.slice.call(tabContents);
-      const item = parent?.querySelectorAll(TARGET + "_list li button");
-      const itemArr = Array.prototype.slice.call(item);
-      const index = itemArr.indexOf(e.target);
-      itemArr.forEach((el) => {
-        el.classList.remove(OPEN);
-        el.setAttribute("aria-pressed", "false");
-        el.setAttribute("tabindex", "0");
-      });
-      event.classList.add(OPEN);
-      event.setAttribute("aria-pressed", "true");
-      event.setAttribute("tabindex", "-1");
-      tabArr.forEach((tab) => {
-        tab.setAttribute("hidden", "");
-        tab.setAttribute("tabindex", "-1");
-      });
-      tabArr[index].removeAttribute("hidden");
-      tabArr[index].focus({ preventScroll: true });
-    }
-    btns.forEach((btn) => {
-      btn.addEventListener("click", onTabClick);
-    });
-    if (btns.length > 0) {
-      const url = new URL(window.location.href);
-      const hash = url.hash;
-      if (hash) {
-        const number = Number(hash.slice(-1));
-        if (!isNaN(number)) {
-          btns[number - 1].click();
-        }
+    const nav = document.querySelector(TARGET);
+    const btn = nav?.querySelector(".ac_menu");
+    const btn_label = nav?.querySelector(".ac_menu span");
+    const wrap = nav?.querySelector(".naviWrapper");
+    const close_btn = nav?.querySelector(".closeBtn");
+    btn?.addEventListener("click", () => {
+      nav?.classList.toggle(OPEN);
+      if (nav?.classList.contains(OPEN)) {
+        btn.setAttribute("aria-expanded", "true");
+        btn_label.textContent = "メニューを閉じる";
+      } else {
+        btn.setAttribute("aria-expanded", "false");
+        btn_label.textContent = "メニューを開く";
       }
-    }
+    });
+    close_btn?.addEventListener("click", () => {
+      this.menuClose(nav, btn, btn_label, OPEN);
+    });
+    wrap?.addEventListener("click", (e) => {
+      const target = e.target;
+      if (target.closest("#navi") === null) {
+        this.menuClose(nav, btn, btn_label, OPEN);
+      }
+    });
+    const subOpen = nav?.querySelectorAll(".spAccordion");
+    let openFlg = false;
+    subOpen?.forEach((btn2) => {
+      btn2?.addEventListener("click", () => {
+        const hasChild = btn2.closest(".has-child");
+        if (!openFlg) {
+          hasChild.classList.add(OPEN);
+          openFlg = true;
+        } else if (openFlg) {
+          hasChild.classList.remove(OPEN);
+          openFlg = false;
+        }
+      });
+    });
+  }
+  /**
+   * メニューを閉じる処理
+   * @param {Element | null} nav - ナビゲーション要素
+   * @param {HTMLElement | null} btn - ハンバーガーボタン要素
+   * @param {HTMLElement | null} btn_label - ハンバーガーボタン内のテキスト要素
+   * @param {string} openClass - メニュー展開時に付与するクラス名
+   * @private
+   */
+  menuClose(nav, btn, btn_label, openClass) {
+    nav?.classList.remove(openClass);
+    btn?.setAttribute("aria-expanded", "false");
+    btn_label.textContent = "メニューを開く";
   }
 }
 
@@ -336,12 +171,7 @@ class SmoothScroll {
 }
 
 window.addEventListener("DOMContentLoaded", () => {
-  new Modal(); // モーダル
-  new Tab(); // タブ
+  new Hamburger(); // ハンバーガー
   new ScrollAnimation(); // アニメーション
   new SmoothScroll(); // スムーススクロール
-});
-
-window.addEventListener("load", () => {
-  new Accordion(); // アコーディオン
 });
