+++ htdocs/sample/index.html
@@ -1,36 +1,37 @@
 <!DOCTYPE html>
 <html dir="ltr" lang="ja">
-  <head prefix="og: http://ogp.me/ns# fb: http://ogp.me/ns/fb# website: http://ogp.me/ns/website#">
+  <head prefix="og: http://ogp.me/ns# fb: http://ogp.me/ns/fb# http://ogp.me/ns/fb# article: http://ogp.me/ns/article#">
 	<meta charset="utf-8">
   <!-- GTAG -->
 	<meta http-equiv="X-UA-Compatible" content="IE=edge">
 	<meta name="viewport" content="width=device-width,initial-scale=1.0,shrink-to-fit=no">
 	<meta name="format-detection' content='telephone=no">
-	<title>タイトルタイトルタイトル</title>
+	<title>コンポーネント｜XXXXX</title>
 	<meta name="description" content="テキストテキストテキストテキストテキストテキスト">
-	<meta property="og:type" content="website">
+	<meta property="og:type" content="article">
 	<meta property="og:local" content="ja_JP">
 	<meta property="og:site_name" content="◯◯◯◯◯◯">
-	<meta property="og:title" content="タイトルタイトルタイトル">
+	<meta property="og:title" content="コンポーネント｜XXXXX">
 	<meta property="og:description" content="テキストテキストテキストテキストテキストテキスト">
-	<meta property="og:url" content="https://www.XXXXX.XX.XX/">
+	<meta property="og:url" content="https://www.XXXXX.XX.XX/sample">
 	<meta property="og:image" content="https://www.XXXXX.XX.XX/_assets/img/ogimg.jpg">
   
 	<meta name="twitter:card" content="summary_large_image">
-	<meta name="twitter:title" content="タイトルタイトルタイトル">
+	<meta name="twitter:title" content="コンポーネント｜XXXXX">
 	<meta name="twitter:description" content="テキストテキストテキストテキストテキストテキスト">
 	<meta name="twitter:image" content="https://www.XXXXX.XX.XX/_assets/img/ogimg.jpg">
   
 	<link rel="icon" href="/favicon.ico">
   
   
-	<link rel="canonical" href="https://www.XXXXX.XX.XX/">
+	<link rel="canonical" href="https://www.XXXXX.XX.XX/sample">
   <link rel="preconnect" href="https://fonts.googleapis.com">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Noto+Sans+JP:wght@100..900&display=swap">
 <link rel="stylesheet" href="/_assets/css/about-index.css">
-<link rel="stylesheet" href="/_assets/css/index.css"></head>
-  <body class="page-top">
+<link rel="stylesheet" href="/_assets/css/about-index-1.css">
+<link rel="stylesheet" href="/_assets/css/sample-index.css"></head>
+  <body class="page-sample">
     <noscript>
   <p>
     このページは JavaScript を使用しています。<br>お使いのブラウザーは
@@ -45,9 +46,9 @@
   <div class="contentInner" data-astro-cid-xbstl6g3>
     <div class="flex items-center justify-between" data-astro-cid-xbstl6g3>
       <a href="/" data-astro-cid-xbstl6g3>
-        <h1 class="headerLogo" data-astro-cid-xbstl6g3>
+        <div class="headerLogo" data-astro-cid-xbstl6g3>
               <img src="/_assets/img/logo_Z1NoFHk.svg" alt="◯◯◯◯◯◯" width="80" height="80" loading="eager" decoding="async">
-            </h1>
+            </div>
       </a>
       <nav class="headerNavi" data-astro-cid-ftm5l6bh>
   <h2 class="txtHidden" data-astro-cid-ftm5l6bh>グローバルナビゲーション</h2>
@@ -97,72 +98,41 @@
 </header>
     <main id="main">
       
-  <div class="top_mv">
-  <div class="c_slide">
-    <div class="splide mvSlide">
-      <div class="c_slide_arrow">
-        <div class="splide__track">
-          <div class="splide__list">
-            <div class="splide__slide">
-                  <picture>
-			<source media="(min-width: 768px)" srcset="/_assets/img/mv_Z1T8rs9.jpg" width="1366" height="518">
-			<img src="/_assets/img/mv_sp_8UYi4.jpg" alt width="375" height="480" class="w-full" loading="lazy" decoding="async">
-		</picture>
-                </div><div class="splide__slide">
-                  <picture>
-			<source media="(min-width: 768px)" srcset="/_assets/img/mv_Z1T8rs9.jpg" width="1366" height="518">
-			<img src="/_assets/img/mv_sp_8UYi4.jpg" alt width="375" height="480" class="w-full" loading="lazy" decoding="async">
-		</picture>
-                </div><div class="splide__slide">
-                  <picture>
-			<source media="(min-width: 768px)" srcset="/_assets/img/mv_Z1T8rs9.jpg" width="1366" height="518">
-			<img src="/_assets/img/mv_sp_8UYi4.jpg" alt width="375" height="480" class="w-full" loading="lazy" decoding="async">
-		</picture>
-                </div>
-          </div>
-        </div>
-      </div>
-      <div class="c_slide_controller">
-        <div class="c_slide_controller_grid">
-          <div class="c_slide_counter">
-            <span class="c_slide_counter_current"></span>
-            <span class="c_slide_counter_length"></span>
-          </div>
-          <div class="c_slide_bar_wrap">
-            <div class="c_slide_bar"></div>
-          </div>
-        </div>
-        <button class="splide__toggle" type="button">
-          <span class="splide__toggle__play">再生</span>
-          <span class="splide__toggle__pause">一時停止</span>
-        </button>
-      </div>
-    </div>
+  <div class="bg-highlight" data-astro-cid-k5h775l4>
+  <div class="contentInner" data-astro-cid-k5h775l4>
+    <h1 class="c_ttl_lower" data-astro-cid-k5h775l4>コンポーネント</h1>
   </div>
 </div>
-
-<script type="module" src="/_assets/js/app2.js"></script>
+  <div class="mb-40" data-astro-cid-ilhxcym7>
+  <div class="contentInner" data-astro-cid-ilhxcym7>
+    <nav id="breadcrumbs" data-astro-cid-ilhxcym7>
+      <ol class="flex gap-30 md:gap-34" itemscope itemtype="https://schema.org/BreadcrumbList" data-astro-cid-ilhxcym7>
+        <li itemprop="itemListElement" itemscope itemtype="https://schema.org/ListItem" data-astro-cid-ilhxcym7>
+              <a href="/" itemprop="item" data-astro-cid-ilhxcym7>
+                  <span itemprop="name" data-astro-cid-ilhxcym7>TOP</span>
+                </a>
+              <meta itemprop="position" content="1">
+            </li><li itemprop="itemListElement" itemscope itemtype="https://schema.org/ListItem" data-astro-cid-ilhxcym7>
+              <span itemprop="name" data-astro-cid-ilhxcym7>コンポーネント</span>
+              <meta itemprop="position" content="2">
+            </li>
+      </ol>
+    </nav>
+  </div>
+</div>
   <div class="grid gap-60">
-    <section>
-  <div class="contentInner scrollIn -up">
-    <h2 class="text-h2 mb-24 font-bold">Tailwind4.0アップデートによる変更点</h2>
-    <div class="mb-64 grid gap-64">
-      <div>
-        <h3 class="text-h3 mb-24 font-bold">
-          カスタムクラスをcssで設定できるように！
-        </h3>
-        <p>
-          案件ごとのカスタムクラスはtailwind.config.jsに記述する必要がありましたが、cssファイルに@themeで設定できるようになりました。<br>また設定したものがcss変数としてデフォルトで出力されるので、sassやjsでも同じ変数を使えるようになりました。
-        </p>
-        <div class="grid grid-cols-2 gap-28">
-          <picture>
-			<source media="(min-width: 768px)" srcset="/_assets/img/image_1ew4eN.png">
-			<img src="/_assets/img/image_sp_Z2nKB9r.png" alt loading="lazy" decoding="async">
-		</picture>
-          <img src="/_assets/img/top/sample.png" alt="">
-        </div>
-      </div>
-    </div>
+    <section data-astro-cid-6sqfndmf>
+  <div class="contentInner" data-astro-cid-6sqfndmf>
+    <h2 class="text-h2 mb-28" data-astro-cid-6sqfndmf>コンポーネントリスト</h2>
+    <ul data-astro-cid-6sqfndmf>
+      <li data-astro-cid-6sqfndmf>
+            <a href="tab" data-astro-cid-6sqfndmf>タブコンポーネント</a>
+          </li><li data-astro-cid-6sqfndmf>
+            <a href="accordion" data-astro-cid-6sqfndmf>アコーディオンコンポーネント</a>
+          </li><li data-astro-cid-6sqfndmf>
+            <a href="modal" data-astro-cid-6sqfndmf>モーダルコンポーネント</a>
+          </li>
+    </ul>
   </div>
 </section>
   </div>
