+++ htdocs/sample/tab/index.html
@@ -6,31 +6,32 @@
 	<meta http-equiv="X-UA-Compatible" content="IE=edge">
 	<meta name="viewport" content="width=device-width,initial-scale=1.0,shrink-to-fit=no">
 	<meta name="format-detection' content='telephone=no">
-	<title>XXXについて｜XXXXX</title>
+	<title>タブコンポーネント｜コンポーネント｜XXX</title>
 	<meta name="description" content="テキストテキストテキストテキストテキストテキスト">
 	<meta property="og:type" content="article">
 	<meta property="og:local" content="ja_JP">
 	<meta property="og:site_name" content="◯◯◯◯◯◯">
-	<meta property="og:title" content="XXXについて｜XXXXX">
+	<meta property="og:title" content="タブコンポーネント｜コンポーネント｜XXX">
 	<meta property="og:description" content="テキストテキストテキストテキストテキストテキスト">
-	<meta property="og:url" content="https://www.XXXXX.XX.XX/">
+	<meta property="og:url" content="https://www.XXXXX.XX.XX/component/tab/">
 	<meta property="og:image" content="https://www.XXXXX.XX.XX/_assets/img/ogimg.jpg">
   
 	<meta name="twitter:card" content="summary_large_image">
-	<meta name="twitter:title" content="XXXについて｜XXXXX">
+	<meta name="twitter:title" content="タブコンポーネント｜コンポーネント｜XXX">
 	<meta name="twitter:description" content="テキストテキストテキストテキストテキストテキスト">
 	<meta name="twitter:image" content="https://www.XXXXX.XX.XX/_assets/img/ogimg.jpg">
   
 	<link rel="icon" href="/favicon.ico">
   
   
-	<link rel="canonical" href="https://www.XXXXX.XX.XX/">
+	<link rel="canonical" href="https://www.XXXXX.XX.XX/component/tab/">
   <link rel="preconnect" href="https://fonts.googleapis.com">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Noto+Sans+JP:wght@100..900&display=swap">
 <link rel="stylesheet" href="/_assets/css/about-index.css">
-<link rel="stylesheet" href="/_assets/css/about-index-1.css"></head>
-  <body class="page-about">
+<link rel="stylesheet" href="/_assets/css/about-index-1.css">
+<link rel="stylesheet" href="/_assets/css/tab-index.css"></head>
+  <body class="page-tab">
     <noscript>
   <p>
     このページは JavaScript を使用しています。<br>お使いのブラウザーは
@@ -99,7 +100,7 @@
       
   <div class="bg-highlight" data-astro-cid-k5h775l4>
   <div class="contentInner" data-astro-cid-k5h775l4>
-    <h1 class="c_ttl_lower" data-astro-cid-k5h775l4>XXXについて</h1>
+    <h1 class="c_ttl_lower" data-astro-cid-k5h775l4>タブコンポーネント</h1>
   </div>
 </div>
   <div class="mb-40" data-astro-cid-ilhxcym7>
@@ -112,13 +113,54 @@
                 </a>
               <meta itemprop="position" content="1">
             </li><li itemprop="itemListElement" itemscope itemtype="https://schema.org/ListItem" data-astro-cid-ilhxcym7>
-              <span itemprop="name" data-astro-cid-ilhxcym7>XXXについて</span>
+              <a href="about" itemprop="item" data-astro-cid-ilhxcym7>
+                  <span itemprop="name" data-astro-cid-ilhxcym7>コンポーネント</span>
+                </a>
               <meta itemprop="position" content="2">
+            </li><li itemprop="itemListElement" itemscope itemtype="https://schema.org/ListItem" data-astro-cid-ilhxcym7>
+              <span itemprop="name" data-astro-cid-ilhxcym7>タブコンポーネント</span>
+              <meta itemprop="position" content="3">
             </li>
       </ol>
     </nav>
   </div>
 </div>
+  <div class="grid gap-60">
+    <section data-astro-cid-n5j6xton>
+  <div class="contentInner" data-astro-cid-n5j6xton>
+    <h2 class="text-h2 text-prime font-bold md:mb-28" data-astro-cid-n5j6xton>タブコンポーネント</h2>
+    <div class="c_tab" data-astro-cid-n5j6xton>
+      <ul class="c_tab_list" data-astro-cid-n5j6xton>
+        <li data-astro-cid-n5j6xton>
+              <button id="tab-1" aria-controls="tabpanel-1" class="-open" aria-pressed="true" tabindex="-1" data-astro-cid-n5j6xton>
+                タブA
+              </button>
+            </li><li data-astro-cid-n5j6xton>
+              <button id="tab-2" aria-controls="tabpanel-2" aria-pressed="false" tabindex="0" data-astro-cid-n5j6xton>
+                タブB
+              </button>
+            </li><li data-astro-cid-n5j6xton>
+              <button id="tab-3" aria-controls="tabpanel-3" aria-pressed="false" tabindex="0" data-astro-cid-n5j6xton>
+                タブC
+              </button>
+            </li>
+      </ul>
+      <div class="c_tab_content" role="tabpanel" id="tabpanel1" aria-labelledby="tab1" data-astro-cid-n5j6xton>
+            <h3 class="text-h3 mb-16 font-bold" data-astro-cid-n5j6xton>タブAの中身</h3>
+            <p data-astro-cid-n5j6xton>あのイーハトーヴォのすきとおった風、夏でも底に冷たさをもつ青いそら、うつくしい森で飾られたモリーオ市、郊外のぎらぎらひかる草の波。<br>テキストテキストテキスト</p>
+          </div><div class="c_tab_content" role="tabpanel" id="tabpanel2" aria-labelledby="tab2" hidden data-astro-cid-n5j6xton>
+            <h3 class="text-h3 mb-16 font-bold" data-astro-cid-n5j6xton>タブBの中身</h3>
+            <p data-astro-cid-n5j6xton>テキストテキストテキストテキストテキストテキストテキストテキスト</p>
+          </div><div class="c_tab_content" role="tabpanel" id="tabpanel3" aria-labelledby="tab3" hidden data-astro-cid-n5j6xton>
+            <h3 class="text-h3 mb-16 font-bold" data-astro-cid-n5j6xton>タブCの中身</h3>
+            <p data-astro-cid-n5j6xton>あのイーハトーヴォのすきとおった風、夏でも底に冷たさをもつ青いそら、うつくしい森で飾られたモリーオ市、郊外のぎらぎらひかる草の波。<br>あのイーハトーヴォのすきとおった風、夏でも底に冷たさをもつ青いそら、うつくしい森で飾られたモリーオ市、郊外のぎらぎらひかる草の波。</p>
+          </div>
+    </div>
+  </div>
+</section>
+
+<script type="module" src="/_assets/js/app3.js"></script>
+  </div>
 
     </main>
     <footer>
