+++ src/components/LowerTitle.astro
@@ -18,7 +18,7 @@ const { title } = Astro.props;
   </div>
 </div>
 
-<style lang="scss">
+<style lang="scss" is:global>
   .c_ttl_lower {
     font-size: $text-h1;
     text-align: center;
