+++ src/layouts/Gnavi.astro
@@ -359,10 +359,3 @@ const { menu, domain } = Astro.props;
 		}
 	}
 </style>
-
-<script>
-  import { Hamburger } from '../ts/class/Hamburger';
-  window.addEventListener('DOMContentLoaded', () => {
-	new Hamburger(); // ハンバーガー
-});
-</script>
