--- src/pages/_parts/_tab.astro
@@ -1,38 +1,0 @@
----
-interface Props {
-  tab: {
-    tabList: string;
-    tabPanel: {
-      ttl: string;
-      txt: string;
-    };
-  }[];
-}
-
-const { tab } = Astro.props;
----
-
-<section>
-  <div class="contentInner">
-    <h2 class="text-h2 text-prime font-bold md:mb-28">タブコンポーネント</h2>
-    <div class="c_tab">
-      <ul class="c_tab_list">
-        {
-          tab.map((item, i) => (
-            <li>
-              <button
-                id={"tab-" + (i + 1)}
-                aria-controls={"tabpanel-" + (i + 1)}
-                class={i === 0 ? "-open" : null}
-                aria-pressed={i === 0 ? "true" : "false"}
-                tabindex={i === 0 ? "-1" : "0"}
-              >
-                {item.tabList}
-              </button>
-            </li>
-          ))
-        }
-      </ul>
-    </div>
-  </div>
-</section>
