+++ src/pages/sample/tab/index.astro
@@ -2,7 +2,7 @@
 import Layout from "@/layouts/Layout.astro";
 import LowerTitle from "@/components/LowerTitle.astro";
 import Breadcrumbs from "@/components/Breadcrumbs.astro";
-import Tab from "@/pages/_parts/_tab.astro";
+import Tab from "@/pages/sample/tab/_parts/_tab.astro";
 
 const page = {
   head: {
@@ -46,7 +46,7 @@ const page = {
         tabList: "タブC",
         tabPanel: {
           ttl: "タブCの中身",
-          txt: "あのイーハトーヴォのすきとおった風、夏でも底に冷たさをもつ青いそら、うつくしい森で飾られたモリーオ市、郊外のぎらぎらひかる草の波。",
+          txt: "あのイーハトーヴォのすきとおった風、夏でも底に冷たさをもつ青いそら、うつくしい森で飾られたモリーオ市、郊外のぎらぎらひかる草の波。<br>あのイーハトーヴォのすきとおった風、夏でも底に冷たさをもつ青いそら、うつくしい森で飾られたモリーオ市、郊外のぎらぎらひかる草の波。",
         },
       },
     ],
